+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add all rooms to the `CustomEngine` class.
 - Add room name prefixes to object base classes.
 - Allow the specifying of music fade times in `CustomEngine` classes.
+- Fix a bug where surface events were not added.
 
 ## 0.4.0